Conform ArgumentsArray to Equatable;
