+++ Sources/Commandant/ArgumentProtocol.swift
@@ -31,27 +31,6 @@ extension String: ArgumentProtocol {
 	}
 }
 
-/// Represents an array of arguments that are converted from the given string
-/// Using comma as a separator
-/// E.g. "banana, apple, peach" -> ["banana", "apple", "peach"]
-struct ArgumentsArray: ArgumentProtocol {
-	/// Array of arguments that were passed by a comma separated string
-	public let arguments: Array<String>
-	
-	/// Separator to be used to split string into values
-	private static let separator = ","
-	
-	public static let name = "array"
-	
-	public static func from(string: String) -> ArgumentsArray? {
-		let trimmedComponents = string.components(separatedBy: separator).map { $0.trimmingCharacters(in: .whitespacesAndNewlines) }
-		if trimmedComponents.count > 1 {
-			return ArgumentsArray(arguments: trimmedComponents)
-		}
-		return string.rangeOfCharacter(from: .whitespaces) == nil ? ArgumentsArray(arguments: [string]) : nil
-	}
-}
-
 // MARK: - migration support
 @available(*, unavailable, renamed: "ArgumentProtocol")
 public typealias ArgumentType = ArgumentProtocol
